Renamed package and modified import statements in test files
