+++ probabilityestimation/probability/.ipynb_checkpoints/exponential-checkpoint.py
