+++ probabilityestimation/probability/gamma.py
